# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateRequest:
    properties:
      cpu:
        description: vCPU 개수
        example: 4
        type: integer
      gfs_mount_point:
        description: GFS mount point. ablestack-vm에서 ccvm.qcow2 경로 계산에 사용한다.
        example: /mnt/glue-gfs
        type: string
      management_network_bridge:
        description: management network bridge
        example: br0
        type: string
      memory:
        description: 메모리 크기(GiB)
        example: 16
        type: integer
      service_network_bridge:
        description: optional service network bridge
        example: br1
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateResponse:
    properties:
      code:
        description: 처리 결과 코드
        example: 200
        type: integer
      message:
        description: 처리 결과 메시지
        example: 클라우드센터 가상머신 xml 생성 성공
        type: string
      results:
        description: fan-out 대상별 결과
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyResult'
        type: array
      val:
        description: 기존 Python createReturn 호환 결과값
      xml_path:
        description: 생성된 XML 파일 경로
        example: /etc/ablestack/vmconfig/ccvm/ccvm.xml
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyResult:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      target:
        example: 10.10.31.1
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.GFSResourceStatusResponse:
    properties:
      code:
        example: 200
        type: integer
      val: {}
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/ccvm/xml:
    post:
      consumes:
      - application/json
      description: Cloud Center VM XML을 생성하고 cluster.json 대상에 배포합니다.
      parameters:
      - description: ccvm xml create request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create CCVM XML
      tags:
      - Cube-CCVM
...
  /cube/gfs/resource/status:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: pcs status xml 기반 펜싱, 잠금 및 glue-gfs/glue-gfs_res GFS2 마운트 리소스 상태를
        조회합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GFSResourceStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.GFSResourceStatusResponse'
      summary: GFS Resource Status
      tags:
      - Cube-GFS
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"